start working on the parser
syntax::Node is an enumeration of possible syntax tree nodes, with None
being the default state. Ident contains a token (TODO: replace with
TokenIndex), which corresponds to the token that forms the identifier.

parse::Parser represents parser context and state. Contains reference to
token stream, reference to source code (so we can resolve tokens into
strings), and "tree" of nodes, which is just a flat array.

We expose parse::Parser::parse_def to try and parse `def` declarations
from the token stream the parser is working on. It can almost do this,
we just need few methods to make it fully work.

Signed-off-by: Aodhnait Étaín <aodhneine@pm.me>